IncidentServer/911-location-server-old/requirements.txt
Babel==2.3.4
Cheetah==2.4.4
Flask==0.10.1
Flask-Babel==0.10.0
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
Flask-WhooshAlchemy==0.56
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.3.0
PyMySQL==0.7.2
PyYAML==3.10
SQLAlchemy==1.0.12
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.11.9
Whoosh==2.7.4
alembic==0.8.6
argparse==1.2.1
blinker==1.4
boto==2.40.0
bz2file==0.98
colorama==0.2.5
configobj==4.7.2
decorator==3.4.0
html5lib==0.999
httpretty==0.8.10
itsdangerous==0.24
jsonpatch==1.3
jsonpointer==1.0
numpy==1.11.0
oauth==1.0.1
pbr==1.9.1
prettytable==0.7.2
pyOpenSSL==0.13
pycurl==7.19.3
pyserial==2.6
python-editor==1.0
pytz==2016.4
requests==2.8.1
scikit-learn==0.17.1
scipy==0.13.3
six==1.10.0
smart-open==1.3.2
speaklater==1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
ssh-import-id==3.21
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
yolk==0.4.3
zope.interface==4.0.5
